pyproject.toml
[tool.poetry]
name = "llm"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9,<3.12"
flask = "^3.0.0"
transformers = {extras = ["torch"], version = "^4.35.2"}
julius = "^0.2.7"
pyannote-audio = "^3.1.1"
google-cloud-speech = "^2.23.0"
cython = "^3.0.7"
tqdm = "^4.66.1"
ipywidgets = "^8.1.1"
faiss-cpu = "^1.7.4"
sentence-transformers = "^2.2.2"
google-api-python-client = "^2.115.0"
google-cloud-bigquery-storage = "^2.24.0"
python-dotenv = "^1.0.1"
google-cloud-bigquery = {extras = ["pandas"], version = "^3.17.1"}
pandas-gbq = "^0.20.0"

[tool.poetry.dev-dependencies]
# Add your development dependencies here
jupyterlab = "^4.0.9"
google-api-python-client = "^2.110.0"
pytube = "^15.0.0"


[tool.poetry.group.dev.dependencies]
speechrecognition = "^3.10.1"
pydub = "^0.25.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
